+++ test/integration/config_property_spec.rb
@@ -11,7 +11,7 @@ describe 'ConfigProperty' do
   describe 'test properties create SSL config' do
     it 'should create a sling folder by default when path node does not exist' do
       config_property = @aem.config_property('someproperty', 'Boolean', true)
-      config_property.create('author', 'org.apache.felix.http')
+      config_property.create('org.apache.felix.http')
     end
 
     it 'should create Apache Felix Jetty Based HTTP Service config property correctly when node exists' do
@@ -23,8 +23,8 @@ describe 'ConfigProperty' do
       node.create('sling:OsgiConfig')
 
       config_property = @aem.config_property('org.apache.felix.https.enable', 'Boolean', true)
-      result = config_property.create('author', 'org.apache.felix.http')
-      expect(result.message).to eq('Set author org.apache.felix.http config Boolean property org.apache.felix.https.enable=true')
+      result = config_property.create('org.apache.felix.http')
+      expect(result.message).to eq('Set org.apache.felix.http config Boolean property org.apache.felix.https.enable=true')
 
       # wait until Jetty finishes restart following org.apache.felix.http config change
       aem = @aem.aem
@@ -42,8 +42,8 @@ describe 'ConfigProperty' do
       node.create('sling:OsgiConfig')
 
       config_property = @aem.config_property('pwdreset.authorizables', 'String[]', %w[admin orchestrator deployer])
-      result = config_property.create('author', 'com.shinesolutions.aem.passwordreset.Activator')
-      expect(result.message).to eq('Set author com.shinesolutions.aem.passwordreset.Activator config String[] property pwdreset.authorizables=["admin", "orchestrator", "deployer"]')
+      result = config_property.create('com.shinesolutions.aem.passwordreset.Activator')
+      expect(result.message).to eq('Set com.shinesolutions.aem.passwordreset.Activator config String[] property pwdreset.authorizables=["admin", "orchestrator", "deployer"]')
     end
 
     it 'should create AEM Health Check Servlet config property correctly when node exists' do
@@ -55,8 +55,8 @@ describe 'ConfigProperty' do
       node.create('sling:OsgiConfig')
 
       config_property = @aem.config_property('bundles.ignored', 'String[]', ['com.day.cq.dam.dam-webdav-support'])
-      result = config_property.create('author', 'com.shinesolutions.healthcheck.hc.impl.ActiveBundleHealthCheck')
-      expect(result.message).to eq('Set author com.shinesolutions.healthcheck.hc.impl.ActiveBundleHealthCheck config String[] property bundles.ignored=["com.day.cq.dam.dam-webdav-support"]')
+      result = config_property.create('com.shinesolutions.healthcheck.hc.impl.ActiveBundleHealthCheck')
+      expect(result.message).to eq('Set com.shinesolutions.healthcheck.hc.impl.ActiveBundleHealthCheck config String[] property bundles.ignored=["com.day.cq.dam.dam-webdav-support"]')
     end
   end
 end