+++ test/unit/resources/config_property_spec.rb
@@ -18,12 +18,11 @@ describe 'ConfigProperty' do
         config_node_name: 'org.apache.felix.http',
         type: 'Boolean',
         value: 'true',
-        run_mode: 'author',
         someproperty: 'true',
         someproperty_type_hint: 'Boolean'
       )
       config_property = RubyAem::Resources::ConfigProperty.new(@mock_client, 'someproperty', 'Boolean', 'true')
-      config_property.create('author', 'org.apache.felix.http')
+      config_property.create('org.apache.felix.http')
     end
 
     it 'should call client with expected parameters when property is a keyword' do
@@ -34,12 +33,11 @@ describe 'ConfigProperty' do
         config_node_name: 'org.apache.sling.jcr.davex.impl.servlets.SlingDavExServlet',
         type: 'String',
         value: '/crx/server',
-        run_mode: 'author',
         _alias: '/crx/server',
         alias_type_hint: 'String'
       )
       config_property = RubyAem::Resources::ConfigProperty.new(@mock_client, 'alias', 'String', '/crx/server')
-      config_property.create('author', 'org.apache.sling.jcr.davex.impl.servlets.SlingDavExServlet')
+      config_property.create('org.apache.sling.jcr.davex.impl.servlets.SlingDavExServlet')
     end
   end
 end